# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Base scaffolding for Go APIs with Gin framework",
        "title": "Simple Arq Golang API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "dev@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/platform-settings/marketplace-fee": {
            "get": {
                "description": "Returns the current marketplace fee percentage.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platform-settings"
                ],
                "summary": "Get marketplace fee",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_platformsettings.MarketplaceFeeResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates the marketplace fee percentage (admin/owner only).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "platform-settings"
                ],
                "summary": "Update marketplace fee",
                "parameters": [
                    {
                        "description": "New fee percentage",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_platformsettings.UpdateMarketplaceFeeRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_platformsettings.MarketplaceFeeResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/simple-arq-golang_cmd_api_domains_apierror.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "simple-arq-golang_cmd_api_domains_apierror.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        },
...
        "simple-arq-golang_cmd_api_domains_platformsettings.MarketplaceFeeResponse": {
            "type": "object",
            "properties": {
                "marketplace_fee_percent": {
                    "type": "number"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "simple-arq-golang_cmd_api_domains_platformsettings.UpdateMarketplaceFeeRequest": {
            "type": "object",
            "required": [
                "marketplace_fee_percent"
            ],
            "properties": {
                "marketplace_fee_percent": {
                    "type": "number",
                    "maximum": 100,
                    "minimum": 0
                }
            }
        },
...
    }
}